sway: Reworked "focus nth child in parent" layout keybindings
Using the "focus nth child in parent" keybindings in nested containers
can be annoying when you are trying to switch to the nth tab but your
direct parent is not the tabbed layout. The new keybindings chose either
the closets or furthest tabbed ancestor and then focus the nth child's
last focused leaf container. Automatically focusing the leaf container
eliminates the need to focus the child (possible multiple times) after
switching to a tab that contains nested containers.