rg-fzf: Better support --null-data by ensuring the first visible match is contained in the fzf item
get_utf8_encoded_codepoint_length was fixed to work with 4-byte UTF-8
codepoints. MAX_DISPLAYED_CODEPOINTS and MAX_DISPLAYED_LINES were
introduced to limit size of the item sent to fzf allowing for large
regions like those often produced by --null-data to not overwhelm fzf.
When truncating the region using MAX_DISPLAYED_CODEPOINTS and
MAX_DISPLAYED_LINES, the resulting region is shifted so that the first
visible match is somewhere in the middle of the new region. Continuation
markers are used to signify that the region was truncated at the
beginning and/or end.